changed int to size_t

+++ 9-binary_tree_height.c
@@ -6,7 +6,7 @@
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int lheight, rheight;
+	size_t lheight, rheight;
 
 	if (tree == NULL || (tree->left == NULL && tree->right == NULL))
 		return (0);
